+++ src/main/java/pl/com/bottega/ecommerce/sales/domain/taxes/calculators/BaseTaxCalculator.java
@@ -1,5 +1,7 @@
-package pl.com.bottega.ecommerce.sales.domain.taxes;
+package pl.com.bottega.ecommerce.sales.domain.taxes.calculators;
 
+import pl.com.bottega.ecommerce.sales.domain.taxes.ITaxCalculator;
+import pl.com.bottega.ecommerce.sales.domain.taxes.Tax;
 import pl.com.bottega.ecommerce.sharedkernel.Money;
 
 public abstract class BaseTaxCalculator implements ITaxCalculator {
