+++ src/main/java/pl/com/bottega/ecommerce/sales/domain/taxes/calculators/StandardTaxCalculator.java
@@ -1,4 +1,4 @@
-package pl.com.bottega.ecommerce.sales.domain.taxes;
+package pl.com.bottega.ecommerce.sales.domain.taxes.calculators;
 
 public class StandardTaxCalculator extends BaseTaxCalculator {
     protected double getTaxRatio() {
